feat(frontend): moved mocked auth integration to a separate package
which allows easier switching between other integrations
