+++ apps/frontend/src/app/[locale]/(auth)/login/page.tsx
@@ -4,14 +4,15 @@ import { setRequestLocale } from 'next-intl/server';
 import { headers } from 'next/headers';
 import { notFound } from 'next/navigation';
 import React from 'react';
-import { auth, signIn } from 'src/auth/auth';
-import { providerMap } from 'src/auth/auth.providers';
 
 import { Toaster } from '@o2s/ui/components/toaster';
 
 import { sdk } from '@/api/sdk';
 
 import { generateSeo } from '@/utils/seo';
+
+import { auth, signIn } from '@/auth';
+import { providerMap } from '@/auth/auth.providers';
 
 import { routing } from '@/i18n/routing';
