+++ apps/frontend/src/app/[locale]/[[...slug]]/page.tsx
@@ -3,7 +3,6 @@ import { setRequestLocale } from 'next-intl/server';
 import { headers } from 'next/headers';
 import { notFound } from 'next/navigation';
 import React from 'react';
-import { auth, signIn } from 'src/auth/auth';
 
 import { Separator } from '@o2s/ui/components/separator';
 import { Toaster } from '@o2s/ui/components/toaster';
@@ -12,6 +11,8 @@ import { Typography } from '@o2s/ui/components/typography';
 import { sdk } from '@/api/sdk';
 
 import { generateSeo } from '@/utils/seo';
+
+import { auth, signIn } from '@/auth';
 
 import { GlobalProvider } from '@/providers/GlobalProvider';
