+++ apps/frontend/src/auth/auth.ts
@@ -1,6 +1,6 @@
 import NextAuth, { NextAuthResult } from 'next-auth';
 
-import { Adapter, jwtCallback, sessionCallback } from './auth.config';
+import { Adapter, jwtCallback, sessionCallback, updateOrganization } from './auth.config';
 import { providers } from './auth.providers';
 
 export const nextAuthResult = NextAuth({
@@ -35,3 +35,4 @@ export const nextAuthResult = NextAuth({
 
 export const { handlers, signIn, signOut } = nextAuthResult;
 export const auth: NextAuthResult['auth'] = nextAuthResult.auth;
+export { updateOrganization };
